Consolidate error_handler.ts
