+++ packages/cli-core/package.json
@@ -23,6 +23,7 @@
     "@inquirer/prompts": "^3.0.0",
     "execa": "^9.5.1",
     "kleur": "^4.1.5",
-    "zod": "^3.22.2"
+    "zod": "^3.22.2",
+    "yargs": "^17.7.2"
   }
 }
