+++ packages/cli-core/src/extract_sub_commands.ts
