+++ packages/cli-core/src/index.ts
@@ -3,3 +3,5 @@ export * from './printer/printer.js';
 export * from './printer.js';
 export { ColorName, colorNames, format, Format } from './format/format.js';
 export * from './package-manager-controller/package_manager_controller_factory.js';
+export * from './error_handler.js';
+export * from './extract_sub_commands.js';
