+++ packages/cli/src/test-utils/command_runner.ts
@@ -1,8 +1,10 @@
 import { Argv } from 'yargs';
 import { AsyncLocalStorage } from 'node:async_hooks';
 import { UsageDataEmitter } from '@aws-amplify/platform-core';
-import { generateCommandFailureHandler } from '../error_handler.js';
-import { extractSubCommands } from '../extract_sub_commands.js';
+import {
+  extractSubCommands,
+  generateCommandFailureHandler,
+} from '@aws-amplify/cli-core';
 
 class OutputInterceptor {
   private output = '';
